refactor: add useStatusText hook
